adjustment to comment count

+++ db/models.js
@@ -14,7 +14,7 @@ exports.fetchArticle = (id) => {
           msg: `Incorrect data type`,
         });
       }
-      return db.query(`SELECT articles.*, COUNT(comments.article_id) AS comment_count FROM articles
+      return db.query(`SELECT articles.*, COUNT(comments.article_id)::INT AS comment_count FROM articles
       LEFT JOIN comments ON comments.article_id = articles.article_id
       WHERE articles.article_id = $1
       GROUP BY articles.article_id;`, [id]).then((result)=>{
